// Backend Projects/hiring-management-backend-development/src/users/enums/user.enum.ts
export enum UserRole {
    superAdmin = "Super-Admin",
    admin = "Admin",
    agency = "Agency",
    shortlister = "Shortlister",
    hr = "HR"
};

export enum UserRoleDisplay {
    admin = "Admin",
    shortlister = "Shortlister",
    hr = "HR"
};

export enum UserStatus {
    approved = 'Approved',
    pending = 'Pending',
    deleted = 'Deleted'
};

export enum UserDesignation {
    jse = 'Junior Software Engineer',
    ase = 'Associate Software Engineer',
    sse = 'Senior Software Engineer',
    pm = 'Project Manager',
    rm = 'Relationship Manager',
    hr = 'HR',
    vp = 'Vice President'
}